+++ eagle/notifications.go
@@ -2,6 +2,7 @@ package eagle
 
 import (
 	"github.com/hacdias/eagle/config"
+	"github.com/hacdias/eagle/logging"
 	"go.uber.org/zap"
 	tb "gopkg.in/tucnak/telebot.v2"
 )
@@ -12,10 +13,10 @@ type Notifications struct {
 	b *tb.Bot
 }
 
-func NewNotifications(c *config.Telegram, log *zap.SugaredLogger) (*Notifications, error) {
+func NewNotifications(c *config.Telegram) (*Notifications, error) {
 	n := &Notifications{
 		Telegram:      c,
-		SugaredLogger: log,
+		SugaredLogger: logging.S().Named("notify"),
 	}
 	b, err := tb.NewBot(tb.Settings{Token: n.Token})
 	if err != nil {
@@ -33,17 +34,17 @@ func (n *Notifications) Notify(msg string) {
 	})
 
 	if err != nil {
-		n.Errorf("could not notify: %s", err)
+		n.Error(err)
 	}
 }
 
-func (n *Notifications) NotifyError(err error) {
-	_, err2 := n.b.Send(&tb.Chat{ID: n.ChatID}, "An error occurred:\n"+err.Error(), &tb.SendOptions{
+func (n *Notifications) NotifyError(not error) {
+	_, err := n.b.Send(&tb.Chat{ID: n.ChatID}, "An error occurred:\n"+not.Error(), &tb.SendOptions{
 		DisableWebPagePreview: true,
 		ParseMode:             tb.ModeDefault,
 	})
 
-	if err2 != nil {
-		n.Errorf("could not notify: %s", err2)
+	if err != nil {
+		n.Error(err)
 	}
 }
