+++ eagle/twitter.go
@@ -3,8 +3,11 @@ package eagle
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"net/http"
+	"net/url"
 	"reflect"
+	"strings"
 	"time"
 
 	"github.com/dghubble/oauth1"
@@ -28,9 +31,9 @@ func NewTwitter(opts *config.Twitter) *Twitter {
 }
 
 func (t *Twitter) Syndicate(entry *Entry) (string, error) {
-	/* status := entry.Content
+	status := entry.Content
 	if len(status) > 280 {
-		status = strings.TrimSpace(status[0:230]) + "... " + entry.Permalink
+		status = strings.TrimSpace(status[0:270-len(entry.Permalink)]) + "... " + entry.Permalink
 	}
 
 	u, err := url.Parse("https://api.twitter.com/1.1/statuses/update.json")
@@ -40,12 +43,18 @@ func (t *Twitter) Syndicate(entry *Entry) (string, error) {
 
 	q := u.Query()
 	q.Set("status", status)
+
 	if entry.Metadata.ReplyTo != nil {
-		if strings.HasSuffix(related, "/") {
-			related = strings.TrimSuffix(related, "/")
+		replyTo, err := url.Parse(entry.Metadata.ReplyTo.URL)
+		if err != nil {
+			return "", err
 		}
-		parts := strings.Split(related, "/")
-		q.Set("in_reply_to_status_id", parts[len(parts)-1])
+
+		user := strings.TrimSuffix(replyTo.Path, "/")
+		user = strings.TrimPrefix(user, "/")
+		parts := strings.Split(user, "/")
+
+		q.Set("in_reply_to_status_id", parts[0])
 		q.Set("auto_populate_reply_metadata", "true")
 		// TODO: add attachment_url for retweet with status
 	}
@@ -75,9 +84,7 @@ func (t *Twitter) Syndicate(entry *Entry) (string, error) {
 		return "", fmt.Errorf("got invalid response: %x", tid)
 	}
 
-	return "https://twitter.com/" + t.User + "/status/" + fmt.Sprint(id), nil */
-
-	return "", nil
+	return "https://twitter.com/" + t.User + "/status/" + fmt.Sprint(id), nil
 }
 
 func (t *Twitter) UserExists(user string) (bool, error) {
